test(memory): move large flate regression out of DocumentIssueFocusTest
