+++ tests/PHPUnit/Integration/DocumentIssueFocusTest.php
@@ -36,8 +36,6 @@
 namespace PHPUnitTests\Integration;
 
 use PHPUnitTests\TestCase;
-use Smalot\PdfParser\Config;
-use Smalot\PdfParser\Document;
 use Smalot\PdfParser\Parser;
 
 /**
@@ -111,17 +109,5 @@ class DocumentIssueFocusTest extends TestCase
         // as well as removing \r line-feeds
         $testSubject = '•†‡…—–ƒ⁄‹›−‰„“”‘’‚™ŁŒŠŸŽıłœšž';
         self::assertStringContainsString($testSubject, $details['Subject']);
-    }
-    /**
-     * @group linux-only
-     */
-    public function testParseFileWithLargeFlateStreams(): void
-    {
-        $config = new Config();
-        $config->setRetainImageContent(false);
-        $config->setDecodeMemoryLimit(8 * 1024 * 1024);
-        $document = (new Parser([], $config))->parseFile($this->rootDir.'/samples/bugs/PullRequest457.pdf');
-
-        self::assertCount(28, $document->getPages());
     }
 }
